refactor(testing): restructure testing utilities into modular files
- Split monolithic testing.go into multiple focused files (config.go, context.go, core.go, db.go, helpers.go, plugin.go, service.go)
- Moved mock service setup functions to dedicated files
- Extracted configuration utilities into config.go
- Created separate files for database testing helpers and plugin registration logic
- Organized API and protocol testing utilities
- Removed redundant imports and improved code organization
- Enhanced test context management and lifecycle handling
