+++ core/testing/operation_handler.go
@@ -1,11 +1,12 @@
 package testing
 
 import (
+	"testing"
+
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
 	"go.lumeweb.com/portal/core"
 	"go.lumeweb.com/portal/db/models"
-	"testing"
 )
 
 // OperationHandlerTest encapsulates common setup and assertion logic for operation handler tests.
