+++ core/testing/protocol_pin_handler.go
@@ -3,9 +3,10 @@ package testing
 import (
 	"context"
 	"errors"
+	"testing"
+
 	"go.lumeweb.com/portal/db/models/data_models"
 	"gorm.io/gorm"
-	"testing"
 )
 
 import (
